refactor(desktop): delete the setup job and the pass-through channel module
Self-review of the previous two commits. Four things were carrying complexity
they had not earned.

1. desktop.yml had a whole `setup` job to resolve the channel. The channel is a
   pure function of the trigger, so it is now a workflow-level `env` expression
   — the same one `concurrency.group` already used. That deletes a job, its
   three outputs, the tag plumbing threaded through `needs.setup`, ~25s of
   serial latency before the build matrix starts, and its Blacksmith minutes.
   `stable` is now unreachable by construction rather than by a runtime guard.
   Precedent for `inputs` in workflow-level env: terraform-apply.yml:103.

2. update-channel.js had become a 23-line module whose only job was to re-export
   resolveChannel from channel.js beside one predicate. Folded isUpdaterSupported
   into channel.js and deleted the module and its test file.

3. DESKTOP_URL_SCHEME was an identity alias of DEFAULT_DESKTOP_SCHEME with zero
   consumers, and DESKTOP_CHANNEL_LABEL was referenced only by the test written
   to justify it. Both deleted.

4. desktopReleaseApiPath and releasesPageFor each rebuilt the
   `desktop-<channel>-latest` string in a different layer. Replaced by one
   desktopReleaseTag(channel) that both derive from, so the tag format — the
   actual contract with desktop.yml's publish job — is written once.

No behavior change. 35 desktop tests and 92 web tests pass, tsc reports no new
errors, eslint is clean on every changed file, and the three channels still
resolve to distinct identities with stable frozen.
